fixed typo meaning that xor Ed,Gd in RM was truncated to short, this screwed linux cpu detection as it masked off the AC bit from eflags! now Ubuntu gets further. Also fixed lss accesses in case of exceeding segment limit

+++ src/org/jpc/emulator/execution/opcodes/vm/xor_Ed_Gd.java
@@ -51,8 +51,8 @@ public class xor_Ed_Gd extends Executable
         Reg op1 = cpu.regs[op1Index];
         Reg op2 = cpu.regs[op2Index];
         cpu.of = cpu.af = cpu.cf = false;
-        cpu.flagResult = (short)(op1.get32() ^ op2.get32());
-        op1.set32((short)cpu.flagResult);
+        cpu.flagResult = (op1.get32() ^ op2.get32());
+        op1.set32(cpu.flagResult);
         cpu.flagStatus = SZP;
         return Branch.None;
     }
